requirements.txt
python-telegram-bot>=20.0.0
python-dotenv
ollama
